+++ CHANGELOG.md
@@ -3,6 +3,8 @@
  * Add a toBeginningOfSentenceCase() method which converts the first character
    of a string to uppercase. It may become more clever about that for locales
    with different conventions over time.
+ * Fixed the use of currency-specific decimal places, which weren't being used
+   if the currency was the default for the locale.
 
 ## 0.13.0
  * Add support for compact number formats ("1.2K") and for significant digits in